rework signing
because address/xpub can be passed in for manual funding, we needed to refactor so signing can happen without HDSigner. In case of XPUB its expected an HDSigner is still setup with seed phrase to be able to sign on a PSBT perhaps created on some server and client signs for it in a stateless way.

Added 2 tests, one for signing with WIF and one for signing with XPUB